pyproject.toml
[project]
name = "notebookllama-improved"
version = "0.5.0"
description = "An improved OSS alternative to NotebookLM using Docling and PostgreSQL"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
  "audioop-lts>=0.2.1",
  "elevenlabs>=2.5.0",
  "fastmcp>=2.9.2",
  "ffprobe>=0.5",
  "llama-index-core>=0.12.44",
  "llama-index-embeddings-openai>=0.3.1",
  "llama-index-embeddings-huggingface-api>=0.3.1",
  "llama-index-llms-openai>=0.4.7",
  "llama-index-observability-otel>=0.1.1",
  "llama-index-tools-mcp>=0.2.5",
  "llama-index-workflows>=1.0.1",
  "llama-index-vector-stores-postgres>=0.3.0",
  "llama-index-readers-docling>=0.1.0",
  "llama-index-node-parser-docling>=0.1.0",
  "docling>=2.51.0",
  "docling-core>=2.11.0",
  "docling-parse>=2.7.0",
  "markdown-analysis>=0.1.5",
  "mypy>=1.16.1",
  "opentelemetry-exporter-otlp-proto-http>=1.34.1",
  "plotly>=6.2.0",
  "pre-commit>=4.2.0",
  "psycopg2-binary>=2.9.10",
  "pydub>=0.25.1",
  "pytest>=8.4.1",
  "pytest-asyncio>=1.0.0",
  "python-dotenv>=1.1.1",
  "pyvis>=0.3.2",
  "randomname>=0.2.1",
  "streamlit>=1.46.1",
  "textual>=3.7.1",
  "pgvector>=0.3.0",
  "sqlalchemy>=2.0.0",
  "httpx>=0.27.0"
]

[tool.mypy]
disable_error_code = ["import-not-found", "import-untyped"]

[tool.pytest.ini_options]
pythonpath = ["src"]
